fixed #12
