+++ proto/build.rs
@@ -8,12 +8,9 @@ fn main() {
                 "src/request_protos/create_license.proto",
                 "src/request_protos/license_activation.proto",
                 "src/request_protos/get_license.proto",
-                "src/response_protos/register_store.proto",
-                "src/response_protos/create_product.proto",
-                "src/response_protos/create_license.proto",
-                "src/response_protos/license_activation.proto",
                 "src/database_protos/store.proto",
-                //"src/database_protos/product.proto",
+                "src/database_protos/product.proto",
+                "src/database_protos/license.proto",
                 ], 
             &["src/"])
         .unwrap();
