+++ proto/src/database_protos/product.proto
@@ -2,15 +2,12 @@ syntax = "proto3";
 
 package product_db_item;
 
-// store database item. This is stored in a protobuf message so that its contents can be easily encrypted. The content in the encrypted message will not be able to be analyzed with tools such as AWS Athena.
+// store database item. This is stored in a protobuf message so that its 
+// contents can be easily encrypted. The content in the encrypted message
+// will not be able to be analyzed with tools such as AWS Athena.
 message ProductDbItem {
     string version = 1;
     bytes store_id = 10;
     bytes product_id = 12;
     string product_name = 20;
-
-    bool is_offline_allowed = 40;
-    
-    uint32 max_machines_per_license = 60;
-
 }
