+++ proto/src/protos/create_license_request.rs
@@ -1,5 +1,4 @@
 // This file is @generated by prost-build.
-/// A store registration request
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct CreateLicenseRequest {
@@ -41,18 +40,25 @@ pub struct ProductInfo {
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct LicenseDbItem {
-    #[prost(bytes = "vec", tag = "10")]
-    pub license_id: ::prost::alloc::vec::Vec<u8>,
-    #[prost(string, tag = "20")]
-    pub customer_first_name: ::prost::alloc::string::String,
-    #[prost(string, tag = "30")]
-    pub customer_last_name: ::prost::alloc::string::String,
-    #[prost(string, tag = "40")]
-    pub customer_email: ::prost::alloc::string::String,
-    /// a string appended to a license code that can allow the user to
-    /// activate an offline license. This "secret" is used because
-    /// offline machines cannot be reliably deactivated
-    #[prost(string, tag = "50")]
-    pub offline_secret: ::prost::alloc::string::String,
+pub struct CreateLicenseResponse {
+    /// this license code will be used for all products from this store for
+    /// this user
+    #[prost(string, tag = "1")]
+    pub license_code: ::prost::alloc::string::String,
+    /// the offline code (if there is one) for the user to activate an offline
+    /// license for their licenses
+    #[prost(string, tag = "2")]
+    pub offline_code: ::prost::alloc::string::String,
+    /// the updated machine limits for each product
+    #[prost(map = "string, uint64", tag = "3")]
+    pub machine_limits: ::std::collections::HashMap<::prost::alloc::string::String, u64>,
+    /// if there are any issues, there will be an error message corresponding to
+    /// the product ID
+    #[prost(map = "string, string", tag = "4")]
+    pub issues: ::std::collections::HashMap<
+        ::prost::alloc::string::String,
+        ::prost::alloc::string::String,
+    >,
+    #[prost(uint64, tag = "10")]
+    pub timestamp: u64,
 }
