--- proto/src/protos/create_license_response.rs
@@ -1,26 +1,0 @@
-// This file is @generated by prost-build.
-/// store registration response
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct CreateLicenseResponse {
-    /// this license code will be used for all products from this store for
-    /// this user
-    #[prost(string, tag = "1")]
-    pub license_code: ::prost::alloc::string::String,
-    /// the offline code (if there is one) for the user to activate an offline
-    /// license for their licenses
-    #[prost(string, tag = "2")]
-    pub offline_code: ::prost::alloc::string::String,
-    /// the updated machine limits for each product
-    #[prost(map = "string, uint64", tag = "3")]
-    pub machine_limits: ::std::collections::HashMap<::prost::alloc::string::String, u64>,
-    /// if there are any issues, there will be an error message corresponding to
-    /// the product ID
-    #[prost(map = "string, string", tag = "4")]
-    pub issues: ::std::collections::HashMap<
-        ::prost::alloc::string::String,
-        ::prost::alloc::string::String,
-    >,
-    #[prost(uint64, tag = "10")]
-    pub timestamp: u64,
-}
