+++ proto/src/protos/create_product_request.rs
@@ -1,5 +1,4 @@
 // This file is @generated by prost-build.
-/// store database item. This is stored in a protobuf message so that its contents can be easily encrypted. The content in the encrypted message will not be able to be analyzed with tools such as AWS Athena.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct CreateProductRequest {
@@ -16,16 +15,17 @@ pub struct CreateProductRequest {
     #[prost(uint64, tag = "150")]
     pub timestamp: u64,
 }
-/// store database item. This is stored in a protobuf message so that its contents can be easily encrypted. The content in the encrypted message will not be able to be analyzed with tools such as AWS Athena.
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ProductDbItem {
+pub struct CreateProductResponse {
+    /// the product id that you will need to specify in the license_auth
+    /// request
     #[prost(string, tag = "1")]
-    pub version: ::prost::alloc::string::String,
-    #[prost(bytes = "vec", tag = "10")]
-    pub store_id: ::prost::alloc::vec::Vec<u8>,
-    #[prost(bytes = "vec", tag = "12")]
-    pub product_id: ::prost::alloc::vec::Vec<u8>,
-    #[prost(string, tag = "20")]
-    pub product_name: ::prost::alloc::string::String,
+    pub product_id: ::prost::alloc::string::String,
+    /// the public verifying key for this product. You will need to verify
+    /// the server's signature on the key file with this public key.
+    #[prost(bytes = "vec", tag = "2")]
+    pub product_public_key: ::prost::alloc::vec::Vec<u8>,
+    #[prost(uint64, tag = "10")]
+    pub timestamp: u64,
 }
