--- proto/src/protos/create_product_response.rs
@@ -1,16 +1,0 @@
-// This file is @generated by prost-build.
-/// store registration response
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct CreateProductResponse {
-    /// the product id that you will need to specify in the license_auth
-    /// request
-    #[prost(string, tag = "1")]
-    pub product_id: ::prost::alloc::string::String,
-    /// the public verifying key for this product. You will need to verify
-    /// the server's signature on the key file with this public key.
-    #[prost(bytes = "vec", tag = "2")]
-    pub product_public_key: ::prost::alloc::vec::Vec<u8>,
-    #[prost(uint64, tag = "10")]
-    pub timestamp: u64,
-}
