+++ proto/src/protos/license_db_item.rs
@@ -1,0 +1,18 @@
+// This file is @generated by prost-build.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct LicenseDbItem {
+    #[prost(bytes = "vec", tag = "10")]
+    pub license_id: ::prost::alloc::vec::Vec<u8>,
+    #[prost(string, tag = "20")]
+    pub customer_first_name: ::prost::alloc::string::String,
+    #[prost(string, tag = "30")]
+    pub customer_last_name: ::prost::alloc::string::String,
+    #[prost(string, tag = "40")]
+    pub customer_email: ::prost::alloc::string::String,
+    /// a string appended to a license code that can allow the user to
+    /// activate an offline license. This "secret" is used because
+    /// offline machines cannot be reliably deactivated
+    #[prost(string, tag = "50")]
+    pub offline_secret: ::prost::alloc::string::String,
+}
