+++ proto/src/protos/mod.rs
@@ -1,15 +1,13 @@
 pub mod register_store_request;
-pub mod register_store_response;
 pub mod store_db_item;
 
 pub mod create_product_request;
-pub mod create_product_response;
+pub mod product_db_item;
 
 pub mod create_license_request;
-pub mod create_license_response;
+pub mod license_db_item;
 
 pub mod license_activation_request;
-pub mod license_activation_response;
 
 pub mod get_license_request;
 
@@ -18,15 +16,13 @@ use zeroize::{Zeroize, ZeroizeOnDrop};
 
 #[cfg(feature = "zeroize")]
 use self::{
-    create_product_request::{CreateProductRequest, ProductDbItem}, 
-    create_product_response::CreateProductResponse, 
-    create_license_request::{CreateLicenseRequest, LicenseDbItem},
-    create_license_response::CreateLicenseResponse,
-    register_store_request::RegisterStoreRequest, 
-    register_store_response::RegisterStoreResponse, 
+    create_product_request::{CreateProductRequest, CreateProductResponse}, 
+    create_license_request::{CreateLicenseRequest, CreateLicenseResponse},
+    register_store_request::{RegisterStoreRequest, RegisterStoreResponse}, 
     store_db_item::StoreDbItem,
-    license_activation_request::{LicenseActivationRequest, Stats},
-    license_activation_response::LicenseKeyFile,
+    license_db_item::LicenseDbItem,
+    product_db_item::ProductDbItem,
+    license_activation_request::{LicenseActivationRequest, Stats, LicenseKeyFile},
     get_license_request::{GetLicenseRequest, GetLicenseResponse},
 };
