+++ proto/src/protos/register_store_request.rs
@@ -65,3 +65,11 @@ pub struct Configs {
     #[prost(uint32, tag = "140")]
     pub trial_license_frequency_hours: u32,
 }
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct RegisterStoreResponse {
+    #[prost(string, tag = "1")]
+    pub store_id: ::prost::alloc::string::String,
+    #[prost(uint64, tag = "10")]
+    pub timestamp: u64,
+}
