+++ proto/src/request_protos/register_store.proto
@@ -49,3 +49,8 @@ message Configs {
     uint32 trial_license_expiration_days = 130;
     uint32 trial_license_frequency_hours = 140;
 }
+
+message RegisterStoreResponse {
+    string store_id = 1;
+    uint64 timestamp = 10;
+}
