--- proto/src/response_protos/create_product.proto
@@ -1,14 +1,0 @@
-syntax = "proto3";
-
-package create_product_response;
-
-// store registration response
-message CreateProductResponse {
-    // the product id that you will need to specify in the license_auth
-    // request
-    string product_id = 1;
-    // the public verifying key for this product. You will need to verify 
-    // the server's signature on the key file with this public key.
-    bytes product_public_key = 2;
-    uint64 timestamp = 10;
-}
