--- proto/src/response_protos/license_activation.proto
@@ -1,40 +1,0 @@
-syntax = "proto3";
-
-package license_activation_response;
-
-// store registration response
-message LicenseKeyFile {
-    string product_id = 1;
-    string customer_first_name = 2;
-    string customer_last_name = 3;
-    string product_version = 4;
-    string license_code = 10;
-    string license_type = 11;
-    string machine_id = 20;
-    uint64 timestamp = 21;
-    // determines when the license expires or needs to be renewed
-    optional uint64 expiration_timestamp = 25;
-    // determines when the client should attempt to renew the license
-    // expiration. Useful for allowing users to deactivate their machines 
-    // that they've stopped using
-    // It is also useful for subscriptions and trials automatically renewing
-    // the expiration
-    optional uint64 check_back_timestamp = 26;
-    string message = 30;
-    // a response code; valid codes are
-    // 1: success
-    // 2: no license found
-    // 4: reached the machine limit
-    // 8: trial ended
-    // 16: license no longer active
-    // 32: incorrect offline code
-    // 64: Offline codes are not allowed for this product
-    uint32 message_code = 31;
-    // the message to show if the license ever expires on the user
-    string post_expiration_message = 35;
-}
-
-message LicenseActivationResponse {
-    LicenseKeyFile key_file = 1;
-    bytes key_file_signature = 2;
-}
